+++ extensions/sovity-messenger/src/test/java/de/sovity/edc/extension/messenger/dto/UnsupportedMessage.java
@@ -5,6 +5,6 @@ import de.sovity.edc.extension.messenger.api.SovityMessage;
 public class UnsupportedMessage implements SovityMessage {
     @Override
     public String getType() {
-        return getClass().getCanonicalName();
+        return "unsupported";
     }
 }